+++ app/controllers/concerns/pundit_helpers.rb
@@ -1,9 +1,9 @@
 module PunditHelpers
   private
 
-  def bulk_authorize(resources)
+  def bulk_authorize(resources, action = nil)
     resources.all? do |resource|
-      authorize(resource)
+      authorize(resource, action)
     end
   end
 end