+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This changelog covers what's changed in the MPDX APIs.
 ## 13 April 2017
 - Added api_class and help_email attributes to a new constant field called organizations_attributes.
 - Added contact referrals relationships to contact serializer.
+- Added people bulk endpoints.
 
 ## 11 April 2017
 - Update Contact suggested_changes after an import.